add short-video ctrlnet model, fix config for sample av post train
