+++ cosmos_transfer1/diffusion/config/training/experiment/ctrl_7b_tp_sample_av.py
@@ -20,10 +20,9 @@ Usage:
 
 import copy
 import os
-from megatron.core import parallel_state
 
 from hydra.core.config_store import ConfigStore
-from torch.utils.data import DataLoader, DistributedSampler
+from torch.utils.data import DataLoader
 from cosmos_transfer1.utils.lazy_config import LazyCall as L
 from cosmos_transfer1.utils.lazy_config import LazyDict
 from cosmos_transfer1.diffusion.config.transfer.conditioner import CTRL_HINT_KEYS_COMB
@@ -33,48 +32,53 @@ from cosmos_transfer1.diffusion.training.networks.general_dit_video_conditioned 
 from cosmos_transfer1.diffusion.training.networks.general_dit import GeneralDIT
 from cosmos_transfer1.diffusion.inference.inference_utils import default_model_names
 from cosmos_transfer1.checkpoints import COSMOS_TRANSFER1_7B_CHECKPOINT, COSMOS_TRANSFER1_7B_SAMPLE_AV_CHECKPOINT
-from cosmos_transfer1.diffusion.datasets.example_transfer_dataset import ExampleTransferDataset, AVTransferDataset
+from cosmos_transfer1.diffusion.datasets.example_transfer_dataset import ExampleTransferDataset
 
 
 cs = ConfigStore.instance()
 
+num_frames = 57
 num_blocks = 28
 num_control_blocks = 3
-ckpt_root = 'checkpoints/'
-data_root = 'datasets/waymo_transfer1/'
+ckpt_root = '/lustre/fsw/portfolios/nvr/users/tianshic/cosmos_ckpts'#"/mnt/scratch/cache/imageinaire/"   #"checkpoints
+data_root = '/lustre/fs12/portfolios/nvr/users/tianshic/jobs/edify_video4/alpamayo_finetune_debug/driving_FT_7Bv312_lvg_1to6_cameras_multi_camera_005_002_frame_repeat_dbg_2_nodes_1_202504231445/cosmos-predict/cosmos-av-sample-toolkits/datasets/waymo_transfer1/'
+
 
 def make_ctrlnet_config(
-    hint_key: str = "control_input_hdmap",
+    hint_key: str = "control_input_segmentation",
     num_control_blocks: int = 3,
     pretrain_model_path: str = "",
-    t2v: bool=True,
-    num_frames=121
+    t2v: bool=True
 ) -> LazyDict:
 
     if pretrain_model_path == "":
         if t2v:
-            job_name = f"CTRL_7Bv1pt3_t2v_{num_frames}frames_{hint_key}_block{num_control_blocks}_pretrain"
+            job_name = f"CTRL_7Bv1pt3_t2v_121frames_{hint_key}_block{num_control_blocks}_pretrain"
             job_project = "cosmos_transfer1_pretrain"
         else:
-            job_name = f"CTRL_7Bv1pt3_lvg_{num_frames}frames_{hint_key}_block{num_control_blocks}_pretrain"
+            job_name = f"CTRL_7Bv1pt3_lvg_121frames_{hint_key}_block{num_control_blocks}_pretrain"
             job_project = "cosmos_transfer1_pretrain"
     else:
         if t2v:
-            job_name = f"CTRL_7Bv1pt3_t2v_{num_frames}frames_{hint_key}_block{num_control_blocks}_posttrain"
+            job_name = f"CTRL_7Bv1pt3_t2v_121frames_{hint_key}_block{num_control_blocks}_posttrain"
             job_project = "cosmos_transfer1_posttrain"
         else:
-            job_name = f"CTRL_7Bv1pt3_lvg_{num_frames}frames_{hint_key}_block{num_control_blocks}_posttrain"
+            job_name = f"CTRL_7Bv1pt3_lvg_121frames_{hint_key}_block{num_control_blocks}_posttrain"
             job_project = "cosmos_transfer1_posttrain"
-    example_multiview_dataset_waymo = L(AVTransferDataset)(
-        dataset_dir=data_root,
-        num_frames=num_frames,
-        hint_key=hint_key,
-        resolution="720",
-        view_keys=["front"],
-        sample_n_views=-1,
-        load_mv_emb=False,
-    )
-
+    # dataset = L(ExampleTransferDataset)(
+    #     "/home/tianshic/code/cosmos-predict1/cosmos-av-sample-toolkits/datasets/waymo/",
+    #     hint_key,
+    #     121,
+    #     "720",
+    #     True
+    # ),
+    # dataset = L(ExampleTransferDataset)(
+    #     dataset_dir="/home/tianshic/code/cosmos-predict1/cosmos-av-sample-toolkits/datasets/waymo/",
+    #     hint_key = hint_key,
+    #     num_frames = 121,
+    #     resolution = "720",
+    #     is_train = True
+    # ),
     ctrl_config = LazyDict(
         dict(
             defaults=[
@@ -86,6 +90,8 @@ def make_ctrlnet_config(
                 {"override /callbacks": "basic"},
                 {"override /checkpoint": "local"},
                 {"override /ckpt_klass": "fast_tp"},
+                {"override /data_train": f"example_transfer_train_data_{hint_key}"},
+                {"override /data_val": f"example_transfer_val_data_{hint_key}"},
                 "_self_",
             ],
             job=dict(group="CTRL_7Bv1_sampleAV", project=job_project, name=job_name),
@@ -154,6 +160,13 @@ def make_ctrlnet_config(
                     pos_emb_learnable=True,
                     extra_per_block_abs_pos_emb_type="learnable",
                 ),
+                # net=L(GeneralDIT)(
+                # net=dict(
+                #     in_channels=16,
+                #     extra_per_block_abs_pos_emb=True,
+                #     pos_emb_learnable=True,
+                #     extra_per_block_abs_pos_emb_type="learnable",
+                # ),
                 adjust_video_noise=True,
                 net_ctrl=dict(
                     in_channels=16,
@@ -165,30 +178,58 @@ def make_ctrlnet_config(
                     extra_per_block_abs_pos_emb_type="learnable",
                 ),
                 tokenizer=dict(
-                    pixel_chunk_duration=num_frames,
+                    #video_vae=dict(
+                        pixel_chunk_duration=num_frames,
+                    #)
                 ),
             ),
             model_obj=L(ShortVideoDiffusionModelWithCtrl)(),
-            dataloader_train=L(DataLoader)(
-                dataset=example_multiview_dataset_waymo,
-                sampler=L(get_sampler)(dataset=example_multiview_dataset_waymo),
-                batch_size=1,
-                drop_last=True,
-                pin_memory=True,
-                num_workers=8
+            # model_obj=L(VideoDiffusionModelWithCtrl)(),
+            #/lustre/fs12/portfolios/nvr/users/tianshic/jobs/edify_video4/alpamayo_finetune_debug/driving_FT_7Bv312_lvg_1to6_cameras_multi_camera_005_002_frame_repeat_dbg_2_nodes_1_202504231445/cosmos-predict/
+            dataloader_train=dict(
+                dataset=dict(dataset_dir=data_root,
+                             num_frames=num_frames
+                             ),
+                sampler=dict(
+                    dataset=dict(dataset_dir=data_root,
+                                 num_frames=num_frames
+                                 )
+                )
             ),
-            dataloader_val=L(DataLoader)(
-                dataset=example_multiview_dataset_waymo,
-                sampler=L(get_sampler)(dataset=example_multiview_dataset_waymo),
-                batch_size=1,
-                drop_last=True,
-                pin_memory=True,
-                num_workers=8
+            dataloader_val=dict(
+                dataset=dict(
+                    dataset_dir=data_root,
+                    num_frames=num_frames
+                ),
+                sampler=dict(
+                    dataset=dict(
+                        dataset_dir=data_root,
+                        num_frames=num_frames
+                    )
+                )
             ),
+
         )
     )
     return ctrl_config
 
+
+def make_small_config(base_config: LazyDict, net_ctrl=False) -> LazyDict:
+    small_config = copy.deepcopy(base_config)
+    small_config["job"]["group"] = "debug"
+    small_config["job"]["name"] = f"{small_config['job']['name']}_SMALL"
+    num_blocks_small = 2
+    small_config["model"]["net"]["num_blocks"] = num_blocks_small
+    if net_ctrl:
+        small_config["model"]["net_ctrl"]["num_blocks"] = num_blocks_small
+        small_config["model"]["net_ctrl"]["layer_mask"] = [
+            True if (i >= num_blocks_small // 2) else False for i in range(num_blocks_small)
+        ]
+    small_config["model_parallel"]["tensor_model_parallel_size"] = 1
+    small_config["model_parallel"]["sequence_parallel"] = False
+    small_config["checkpoint"]["load_path"] = ""
+    small_config["model"]["base_load_from"]["load_path"] = ""
+    return small_config
 
 
 all_hint_key = [
@@ -197,28 +238,43 @@ all_hint_key = [
 ]
 
 for key in all_hint_key:
-    for num_frames in [57, 121]:
-        # Register experiments for pretraining from scratch
-        t2v_config = make_ctrlnet_config(hint_key=key, num_control_blocks=num_control_blocks,
-                                         pretrain_model_path="", t2v=True, num_frames=num_frames)
-        cs.store(
-            group="experiment",
-            package="_global_",
-            name=t2v_config["job"]["name"],
-            node=t2v_config,
-        )
-
-        # Register experiments for post-training from TP checkpoints.
-        hint_key_short = key.replace("control_input_", "")  # "control_input_vis" -> "vis"
-        pretrain_ckpt_path = default_model_names[hint_key_short]
-        # note: The TP ckpt path are specified as <name>.pt to the script, but actually the <name>_model_mp_*.pt files will be loaded.
-        tp_ckpt_path = os.path.join(ckpt_root, os.path.dirname(pretrain_ckpt_path), "checkpoints_tp",
-                                    os.path.basename(pretrain_ckpt_path))
-        config = make_ctrlnet_config(hint_key=key, num_control_blocks=num_control_blocks,
-                                    pretrain_model_path=tp_ckpt_path, t2v=True, num_frames=num_frames)
-        cs.store(
-            group="experiment",
-            package="_global_",
-            name=config["job"]["name"],
-            node=config,
-        )
+    # Register experiments for pretraining from scratch
+    t2v_config = make_ctrlnet_config(hint_key=key, num_control_blocks=num_control_blocks,
+                                     pretrain_model_path="", t2v=True)
+    # never released these lol...
+    i2v_config = make_ctrlnet_config(hint_key=key, num_control_blocks=num_control_blocks,
+                                     pretrain_model_path="", t2v=False)
+    # Register experiments for pretraining from scratch
+    debug_config = make_small_config(t2v_config, net_ctrl=True)
+    cs.store(
+        group="experiment",
+        package="_global_",
+        name=t2v_config["job"]["name"],
+        node=t2v_config,
+    )
+    cs.store(
+        group="experiment",
+        package="_global_",
+        name=i2v_config["job"]["name"],
+        node=i2v_config,
+    )
+    cs.store(
+        group="experiment",
+        package="_global_",
+        name=debug_config["job"]["name"],
+        node=debug_config,
+    )
+    # Register experiments for post-training from TP checkpoints.
+    hint_key_short = key.replace("control_input_", "")  # "control_input_vis" -> "vis"
+    pretrain_ckpt_path = default_model_names[hint_key_short]
+    # note: The TP ckpt path are specified as <name>.pt to the script, but actually the <name>_model_mp_*.pt files will be loaded.
+    tp_ckpt_path = os.path.join(ckpt_root, os.path.dirname(pretrain_ckpt_path), "checkpoints_tp",
+                                os.path.basename(pretrain_ckpt_path))
+    config = make_ctrlnet_config(hint_key=key, num_control_blocks=num_control_blocks,
+                                pretrain_model_path=tp_ckpt_path, t2v=True)
+    cs.store(
+        group="experiment",
+        package="_global_",
+        name=config["job"]["name"],
+        node=config,
+    )
